+++ app/src/main/java/ru/loftschool/moneytrackerloft/MainActivity.java
@@ -1,5 +1,6 @@
 package ru.loftschool.moneytrackerloft;
 
+import android.app.Fragment;
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.view.Menu;
@@ -17,35 +18,15 @@ import java.util.Locale;
 
 public class MainActivity extends ActionBarActivity {
 
-    private ListView listView;
-    private TransactionAdapter transactionAdapter;
-    List<Transaction> data = new ArrayList<>();
 
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        List<Transaction> adapterData = getTransactions();
-        transactionAdapter = new TransactionAdapter(this, adapterData);
-
-        listView = (ListView) findViewById(R.id.listview);
-        listView.setAdapter(transactionAdapter);
+        final TransactionsFragment fragmentById = (TransactionsFragment) getFragmentManager().findFragmentById(R.id.transactions_fragment);
+        fragmentById.setParametr(0);
     }
 
-
-    private String dateFormat = "dd-MM-yyyy";
-
-    private List<Transaction> getTransactions() {
-
-        DateFormat df = new SimpleDateFormat(dateFormat, new Locale("ru"));
-        Date now_calendar = Calendar.getInstance().getTime();
-        String now = df.format(now_calendar);
-        data.add(new Transaction("Huawei", "9800", now));
-        data.add(new Transaction("SamsungS3", "13000", now));
-        data.add(new Transaction("T-shirt", "300", now));
-        data.add(new Transaction("Jeans", "1500", now));
-        data.add(new Transaction("Printer", "4500", now));
-        return data;
-    }
 }
+
