+++ app/src/main/res/layout/fragment_transactions.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+
+    style="@style/Container.Background">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/color_act">
+
+
+        <ListView
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:id="@+id/listview"
+            android:divider="@android:color/transparent"
+            android:dividerHeight="@dimen/dividerHight">
+
+        </ListView>
+    </LinearLayout>
+
+</RelativeLayout>
